This code outputs the prompts.
This code randomly generates a number to call a prompt. This prompt will tell the characters what their new mission is. There are up to 8 prompts for the players because there can only ever be 8 prompts called. This function will be called during new levels, and prompts will be added once game functionality is added. Currently the message is printed in console for debugging.